\name{[[<--methods}
\docType{methods}
\alias{[[<--methods}
\alias{[[<-,Lagged,numeric-method}
\title{Methods for '[[<-' in package 'lagged'}
\description{Methods for '[[<-' in package 'lagged'.}
\section{Methods}{
\describe{

\item{\code{signature(x = "Lagged", i = "numeric")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{[[-methods}}
}
\keyword{methods}
